+++ src/main/java/conflux/web3j/AccountManager.java
@@ -12,6 +12,7 @@ import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
+import conflux.web3j.types.Address;
 import org.web3j.crypto.CipherException;
 import org.web3j.crypto.Credentials;
 import org.web3j.crypto.ECKeyPair;
@@ -24,7 +25,6 @@ import org.web3j.utils.Numeric;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import conflux.web3j.types.Address;
 import conflux.web3j.types.AddressType;
 import conflux.web3j.types.RawTransaction;
 
@@ -42,33 +42,34 @@ public class AccountManager {
 	private String dir;
 	// unlocked accounts: map<address, item>
 	private ConcurrentHashMap<String, UnlockedItem> unlocked;
+
+	private int networkId;
 	
 	/**
 	 * Create a AccountManager instance with default directory.
-	 * @throws IOException if failed to create the default directory.
+	 * @param networkId networkId
+	 * @throws Exception if failed to create directories.
 	 */
-	public AccountManager() {
-		this(getDefaultDirectory());
+	public AccountManager(int networkId) throws Exception {
+		this(getDefaultDirectory(), networkId);
 	}
 	
 	/**
 	 * Create a AccountManager instance with specified directory.
 	 * @param dir directory to store key files.
+	 * @param networkId networkId
 	 * @throws IOException if failed to create directories.
 	 */
-	public AccountManager(String dir) {
-		try {
-			Files.createDirectories(Paths.get(dir));
-		} catch (IOException e) {
-			throw new IllegalArgumentException(e);
-		}
-		
+	public AccountManager(String dir, int networkId) throws IOException {
+		Files.createDirectories(Paths.get(dir));
 		this.dir = dir;
+		this.networkId = networkId;
 		this.unlocked = new ConcurrentHashMap<String, UnlockedItem>();
 	}
 	
 	/**
 	 * Get the directory to store key files.
+	 * @return directory path.
 	 */
 	public String getDirectory() {
 		return this.dir;
@@ -94,12 +95,13 @@ public class AccountManager {
 	 * Create a new account with specified password.
 	 * @param password used to encrypt the key file.
 	 * @return address of new created account.
+	 * @throws Exception if failed to create file
 	 */
-	public String create(String password) throws Exception {
+	public Address create(String password) throws Exception {
 		return this.createKeyFile(password, Keys.createEcKeyPair());
 	}
 	
-	private String createKeyFile(String password, ECKeyPair ecKeyPair) throws Exception {
+	private Address createKeyFile(String password, ECKeyPair ecKeyPair) throws Exception {
 		WalletFile walletFile = Wallet.createStandard(password, ecKeyPair);
 		walletFile.setAddress(AddressType.User.normalize(walletFile.getAddress()));
 
@@ -107,18 +109,20 @@ public class AccountManager {
 		File keyfile = new File(this.dir, filename);
 		objectMapper.writeValue(keyfile, walletFile);
 
-		return walletFile.getAddress();
+		return new Address(walletFile.getAddress(), this.networkId);
 	}
 	
 	/**
 	 * List all managed accounts.
 	 * @return list of addresses of all managed accounts.
+	 * @throws IOException if read files failed
 	 */
-	public List<String> list() throws IOException {
+	public List<Address> list() throws IOException {
 		return Files.list(Paths.get(this.dir))
 				.map(path -> this.parseAddressFromFilename(path.getFileName().toString()))
 				.filter(path -> !path.isEmpty())
 				.sorted()
+				.map(hexAddress -> new Address(hexAddress, this.networkId))
 				.collect(Collectors.toList());
 	}
 	
@@ -132,19 +136,20 @@ public class AccountManager {
 			return "";
 		}
 		
-		String address = filename.substring(keyfilePrefix.length(), filename.length() - keyfileExt.length());
+		String hexAddress = filename.substring(keyfilePrefix.length(), filename.length() - keyfileExt.length());
 		
 		try {
-			Address.validate(address, AddressType.User);
+			AddressType.validateHexAddress(hexAddress, AddressType.User);
 		} catch (Exception e) {
 			return "";
 		}
 		
-		return address;
+		return hexAddress;
 	}
 	
-	private Optional<String> imports(Credentials credentials, String password) throws Exception {
-		String address = AddressType.User.normalize(credentials.getAddress());
+	private Optional<Address> imports(Credentials credentials, String password) throws Exception {
+		String hexAddress = AddressType.User.normalize(credentials.getAddress());
+		Address address = new Address(hexAddress, this.networkId);
 		if (this.exists(address)) {
 			return Optional.empty();
 		}
@@ -160,8 +165,9 @@ public class AccountManager {
 	 * @param password decrypt the external key file.
 	 * @param newPassword encrypt the new created/managed key file.
 	 * @return imported account address if not exists. Otherwise, return <code>Optional.empty()</code>.
+	 * @throws Exception if load keyFile failed
 	 */
-	public Optional<String> imports(String keyFile, String password, String newPassword) throws Exception {
+	public Optional<Address> imports(String keyFile, String password, String newPassword) throws Exception {
 		Credentials importedCredentials = WalletUtils.loadCredentials(password, keyFile);
 		return this.imports(importedCredentials, newPassword);
 	}
@@ -171,8 +177,9 @@ public class AccountManager {
 	 * @param privateKey private key to import.
 	 * @param password encrypt the new created/managed key file.
 	 * @return imported account address if not exists. Otherwise, return <code>Optional.empty()</code>.
+	 * @throws Exception if create file failed
 	 */
-	public Optional<String> imports(String privateKey, String password) throws Exception {
+	public Optional<Address> imports(String privateKey, String password) throws Exception {
 		return this.imports(Credentials.create(privateKey), password);
 	}
 	
@@ -180,11 +187,12 @@ public class AccountManager {
 	 * Check whether the specified account address is managed.
 	 * @param address account address.
 	 * @return <code>true</code> if the specified account address is managed. Otherwise, <code>false</code>.
+	 * @throws Exception if read file failed
 	 */
-	public boolean exists(String address) throws IOException {
+	public boolean exists(Address address) throws Exception {
 		return Files.list(Paths.get(this.dir))
 				.map(path -> this.parseAddressFromFilename(path.getFileName().toString()))
-				.anyMatch(path -> path.equalsIgnoreCase(address));
+				.anyMatch(path -> path.equalsIgnoreCase(address.getHexAddress()));
 	}
 	
 	/**
@@ -192,10 +200,12 @@ public class AccountManager {
 	 * It will also clear the record if the specified account is unlocked.
 	 * @param address account address to delete.
 	 * @return <code>false</code> if the specified account not found. Otherwise, <code>true</code>.
+	 * @throws Exception  if file delete failed
 	 */
-	public boolean delete(String address) throws IOException {
+	public boolean delete(Address address) throws Exception {
+		String hexAddress = address.getHexAddress();
 		List<Path> files = Files.list(Paths.get(this.dir))
-				.filter(path -> this.parseAddressFromFilename(path.getFileName().toString()).equalsIgnoreCase(address))
+				.filter(path -> this.parseAddressFromFilename(path.getFileName().toString()).equalsIgnoreCase(hexAddress))
 				.collect(Collectors.toList());
 		
 		if (files.isEmpty()) {
@@ -206,7 +216,7 @@ public class AccountManager {
 			Files.delete(file);
 		}
 		
-		this.unlocked.remove(address);
+		this.unlocked.remove(hexAddress);
 		
 		return true;
 	}
@@ -217,10 +227,11 @@ public class AccountManager {
 	 * @param password password to decrypt the original key file.
 	 * @param newPassword password to encrypt the new key file.
 	 * @return <code>false</code> if the specified account not found. Otherwise, <code>true</code>.
+	 * @throws Exception if file read failed
 	 */
-	public boolean update(String address, String password, String newPassword) throws Exception {
+	public boolean update(Address address, String password, String newPassword) throws Exception {
 		List<Path> files = Files.list(Paths.get(this.dir))
-				.filter(path -> this.parseAddressFromFilename(path.getFileName().toString()).equalsIgnoreCase(address))
+				.filter(path -> this.parseAddressFromFilename(path.getFileName().toString()).equalsIgnoreCase(address.getHexAddress()))
 				.collect(Collectors.toList());
 		
 		if (files.isEmpty()) {
@@ -239,10 +250,11 @@ public class AccountManager {
 	 * @param address account address to export private key.
 	 * @param password to decrypt the original key file.
 	 * @return private key if account exists. Otherwise, <code>null</code>.
+	 * @throws Exception if file read failed
 	 */
-	public String exportPrivateKey(String address, String password) throws Exception {
+	public String exportPrivateKey(Address address, String password) throws Exception {
 		List<Path> files = Files.list(Paths.get(this.dir))
-				.filter(path -> this.parseAddressFromFilename(path.getFileName().toString()).equalsIgnoreCase(address))
+				.filter(path -> this.parseAddressFromFilename(path.getFileName().toString()).equalsIgnoreCase(address.getHexAddress()))
 				.collect(Collectors.toList());
 		
 		if (files.isEmpty()) {
@@ -259,10 +271,12 @@ public class AccountManager {
 	 * @param password decrypt the key file.
 	 * @param timeout a period of time to unlock the account. Empty timeout indicates unlock the account indefinitely.
 	 * @return <code>false</code> if the specified account not found. Otherwise, <code>true</code>.
+	 * @throws Exception if file read failed
 	 */
-	public boolean unlock(String address, String password, Duration... timeout) throws Exception {
+	public boolean unlock(Address address, String password, Duration... timeout) throws Exception {
+		String hexAddress = address.getHexAddress();
 		List<Path> files = Files.list(Paths.get(this.dir))
-				.filter(path -> this.parseAddressFromFilename(path.getFileName().toString()).equalsIgnoreCase(address))
+				.filter(path -> this.parseAddressFromFilename(path.getFileName().toString()).equalsIgnoreCase(hexAddress))
 				.collect(Collectors.toList());
 		
 		if (files.isEmpty()) {
@@ -279,7 +293,7 @@ public class AccountManager {
 			item = new UnlockedItem(credentials.getEcKeyPair(), Optional.empty());
 		}
 		
-		this.unlocked.put(address, item);
+		this.unlocked.put(hexAddress, item);
 		
 		return true;
 	}
@@ -289,8 +303,8 @@ public class AccountManager {
 	 * @param address account address to lock.
 	 * @return <code>true</code> if the specified account has already been unlocked and not expired. Otherwise, <code>false</code>.
 	 */
-	public boolean lock(String address) {
-		UnlockedItem item = this.unlocked.remove(address);
+	public boolean lock(Address address) {
+		UnlockedItem item = this.unlocked.remove(address.getHexAddress());
 		return item != null && !item.expired();
 	}
 	
@@ -301,13 +315,15 @@ public class AccountManager {
 	 * @param password decrypt the key file. If empty, the account should be unlocked already.
 	 * @return signed and RLP encoded transaction.
 	 * @exception IllegalArgumentException if account not found, or password not specified for locked account, or password expired for unlocked account.
+	 * @throws Exception if get keypair failed
 	 */
-	public String signTransaction(RawTransaction tx, String address, String... password) throws Exception {
+	public String signTransaction(RawTransaction tx, Address address, String... password) throws Exception {
 		ECKeyPair ecKeyPair = this.getEcKeyPair(address, password);
 		return tx.sign(ecKeyPair);
 	}
 	
-	private ECKeyPair getEcKeyPair(String address, String... password) throws IOException, CipherException {
+	private ECKeyPair getEcKeyPair(Address cfxAddress, String... password) throws IOException, CipherException {
+		String address = cfxAddress.getHexAddress();
 		UnlockedItem item = this.unlocked.get(address);
 		
 		if (password == null || password.length == 0 || password[0] == null || password[0].isEmpty()) {	
@@ -342,7 +358,7 @@ public class AccountManager {
 		}
 	}
 	
-	public String signMessage(byte[] message, boolean needToHash, String address, String... password) throws Exception {
+	public String signMessage(byte[] message, boolean needToHash, Address address, String... password) throws Exception {
 		ECKeyPair ecKeyPair = this.getEcKeyPair(address, password);
 		return signMessage(message, needToHash, ecKeyPair);
 	}