+++ src/main/java/conflux/web3j/contract/ContractCall.java
@@ -10,6 +10,7 @@ import org.web3j.abi.datatypes.Type;
 
 import conflux.web3j.Cfx;
 import conflux.web3j.Request;
+import conflux.web3j.types.Address;
 import conflux.web3j.contract.abi.DecodeUtil;
 import conflux.web3j.request.Call;
 import conflux.web3j.request.Epoch;
@@ -23,12 +24,12 @@ public class ContractCall {
 	private Cfx cfx;
 	private Epoch epoch;
 	
-	public ContractCall(Cfx cfx, String contractAddress) {
+	public ContractCall(Cfx cfx, Address contractAddress) {
 		this.cfx = cfx;
 		this.call.setTo(contractAddress);
 	}
 	
-	public ContractCall buildFrom(String from) {
+	public ContractCall buildFrom(Address from) {
 		this.call.setFrom(from);
 		BigInteger nonce = this.epoch == null
 				? this.cfx.getNonce(from).sendAndGet()
@@ -88,7 +89,5 @@ public class ContractCall {
 		String rawData = this.call(method, args).sendAndGet();
 		return DecodeUtil.decode(rawData, returnType);
 	}
-
-
 
 }