+++ src/main/java/conflux/web3j/contract/internals/AdminControl.java
@@ -1,7 +1,6 @@
 package conflux.web3j.contract.internals;
 
-import org.web3j.abi.datatypes.Address;
-
+import conflux.web3j.types.Address;
 import conflux.web3j.Account;
 import conflux.web3j.Account.Option;
 import conflux.web3j.Cfx;
@@ -11,37 +10,40 @@ import conflux.web3j.contract.ContractCall;
 public class AdminControl extends ContractCall {
     private final static String contract = "0x0888000000000000000000000000000000000000";
     private Account account;  // if account not set, can only use getAdmin method
+    private Address contractAddress;
 
-    public AdminControl(Account account) {
-        super(account.getCfx(), AdminControl.contract);
+    public AdminControl(Account account, int networkId) {
+        super(account.getCfx(), new Address(AdminControl.contract, networkId));
         this.account = account;
+        this.contractAddress = new Address(AdminControl.contract, networkId);
     }
 
     public AdminControl(Cfx cfx) {
-        super(cfx, AdminControl.contract);
+        super(cfx, new Address(AdminControl.contract, cfx.getIntNetworkId()));
+        this.contractAddress = new Address(AdminControl.contract, cfx.getIntNetworkId());
     }
 
     public void setAccount(Account account) {
         this.account = account;
     }
 
-    public String getAdmin(String contractAddr) throws RpcException {
-        return this.callAndGet(Address.class, "getAdmin", new Address(contractAddr));
+    public String getAdmin(Address contractAddr) throws RpcException {
+        return this.callAndGet(Address.class, "getAdmin", contractAddr.getABIAddress());
     }
 
-    public String destroy (Option option, String contractAddr) throws Exception {
+    public String destroy (Option option, Address contractAddr) throws Exception {
         String admin = getAdmin(contractAddr);
-        if (!admin.equalsIgnoreCase(account.getAddress())) {
+        if (!admin.equalsIgnoreCase(account.getAddress().getHexAddress())) {
             throw new Exception("Administrator privilege required");
         }
-        return this.account.call(option, contract, "destroy", new Address(contractAddr));
+        return this.account.call(option, this.contractAddress, "destroy", contractAddr.getABIAddress());
     }
 
-    public String setAdmin(Option option, String contractAddr, String newAdmin) throws Exception {
+    public String setAdmin(Option option, Address contractAddr, Address newAdmin) throws Exception {
         String admin = getAdmin(contractAddr);
-        if (!admin.equalsIgnoreCase(account.getAddress())) {
+        if (!admin.equalsIgnoreCase(account.getAddress().getHexAddress())) {
             throw new Exception("Administrator privilege required");
         }
-        return this.account.call(option, contract, "setAdmin", new Address(contractAddr), new Address(newAdmin));
+        return this.account.call(option, this.contractAddress, "setAdmin", contractAddr.getABIAddress(), newAdmin.getABIAddress());
     }
 }