+++ src/main/java/conflux/web3j/contract/internals/SponsorWhitelistControl.java
@@ -5,59 +5,62 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.web3j.abi.datatypes.Address;
 import org.web3j.abi.datatypes.Bool;
 import org.web3j.abi.datatypes.DynamicArray;
 import org.web3j.abi.datatypes.generated.Uint256;
 
 import conflux.web3j.Account;
 import conflux.web3j.Cfx;
+import conflux.web3j.types.Address;
 import conflux.web3j.RpcException;
 import conflux.web3j.contract.ContractCall;
 
 public class SponsorWhitelistControl extends ContractCall {
     private final static String contract = "0x0888000000000000000000000000000000000001";
     private Account account;
+    private Address contractAddress;
 
-    public SponsorWhitelistControl(Account account) {
-        super(account.getCfx(), SponsorWhitelistControl.contract);
+    public SponsorWhitelistControl(Account account, int networkId) {
+        super(account.getCfx(), new Address(SponsorWhitelistControl.contract, networkId));
         this.account = account;
+        this.contractAddress = new Address(SponsorWhitelistControl.contract, networkId);
     }
 
     public SponsorWhitelistControl(Cfx cfx) {
-        super(cfx, SponsorWhitelistControl.contract);
+        super(cfx, new Address(SponsorWhitelistControl.contract, cfx.getIntNetworkId()));
+        this.contractAddress = new Address(SponsorWhitelistControl.contract, cfx.getIntNetworkId());
     }
 
     public void setAccount(Account account) {
         this.account = account;
     }
 
-    public String getSponsorForGas(String contractAddr) throws RpcException {
-        return this.callAndGet(Address.class, "getSponsorForGas", new Address(contractAddr));
+    public String getSponsorForGas(Address contractAddr) throws RpcException {
+        return this.callAndGet(Address.class, "getSponsorForGas", contractAddr.getABIAddress());
     }
 
-    public BigInteger getSponsoredBalanceForGas(String contractAddr) throws RpcException {
-        return this.callAndGet(Uint256.class, "getSponsoredBalanceForGas", new Address(contractAddr));
+    public BigInteger getSponsoredBalanceForGas(Address contractAddr) throws RpcException {
+        return this.callAndGet(Uint256.class, "getSponsoredBalanceForGas", contractAddr.getABIAddress());
     }
 
-    public BigInteger getSponsoredGasFeeUpperBound(String contractAddr) throws RpcException {
-        return this.callAndGet(Uint256.class, "getSponsoredGasFeeUpperBound", new Address(contractAddr));
+    public BigInteger getSponsoredGasFeeUpperBound(Address contractAddr) throws RpcException {
+        return this.callAndGet(Uint256.class, "getSponsoredGasFeeUpperBound", contractAddr.getABIAddress());
     }
 
-    public String getSponsorForCollateral(String contractAddr) throws RpcException {
-        return this.callAndGet(Address.class, "getSponsorForCollateral", new Address(contractAddr));
+    public String getSponsorForCollateral(Address contractAddr) throws RpcException {
+        return this.callAndGet(Address.class, "getSponsorForCollateral", contractAddr.getABIAddress());
     }
 
-    public BigInteger getSponsoredBalanceForCollateral(String contractAddr) throws RpcException {
-        return this.callAndGet(Uint256.class, "getSponsoredBalanceForCollateral", new Address(contractAddr));
+    public BigInteger getSponsoredBalanceForCollateral(Address contractAddr) throws RpcException {
+        return this.callAndGet(Uint256.class, "getSponsoredBalanceForCollateral", contractAddr.getABIAddress());
     }
 
-    public boolean isWhitelisted(String contractAddr, String user) throws RpcException {
-        return this.callAndGet(Bool.class, "isWhitelisted", new Address(contractAddr), new Address(user));
+    public boolean isWhitelisted(Address contractAddr, Address user) throws RpcException {
+        return this.callAndGet(Bool.class, "isWhitelisted", contractAddr.getABIAddress(), contractAddr.getABIAddress());
     }
 
-    public boolean isAllWhitelisted(String contractAddr) throws RpcException {
-        return this.callAndGet(Bool.class, "isAllWhitelisted", new Address(contractAddr));
+    public boolean isAllWhitelisted(Address contractAddr) throws RpcException {
+        return this.callAndGet(Bool.class, "isAllWhitelisted", contractAddr.getABIAddress());
     }
 
 //    public void addPrivilege(Account.Option option, String[] addresses) throws Exception {
@@ -70,21 +73,23 @@ public class SponsorWhitelistControl extends ContractCall {
 //        account.call(option, contract, "removePrivilege", new DynamicArray<Address>(Address.class, list));
 //    }
 
-    public String setSponsorForCollateral(Account.Option option, String contractAddr) throws Exception {
-        return account.call(option, contract, "setSponsorForCollateral", new Address(contractAddr));
+    public String setSponsorForCollateral(Account.Option option, Address contractAddr) throws Exception {
+        return account.call(option, this.contractAddress, "setSponsorForCollateral", contractAddr.getABIAddress());
     }
 
-    public String setSponsorForGas(Account.Option option, String contractAddr, BigInteger upperBound) throws Exception {
-        return account.call(option, contract, "setSponsorForGas", new Address(contractAddr), new Uint256(upperBound));
+    public String setSponsorForGas(Account.Option option, Address contractAddr, BigInteger upperBound) throws Exception {
+        return account.call(option, this.contractAddress, "setSponsorForGas", contractAddr.getABIAddress(), new Uint256(upperBound));
     }
 
-    public String addPrivilegeByAdmin(Account.Option option, String contractAddr, String[] address) throws Exception {
-        List<Address> list = Arrays.stream(address).map(a -> new Address(a)).collect(Collectors.toList());
-        return account.call(option, contract, "addPrivilegeByAdmin", new Address(contractAddr), new DynamicArray<Address>(Address.class, list));
+    public String addPrivilegeByAdmin(Account.Option option, Address contractAddr, Address[] address) throws Exception {
+        List<org.web3j.abi.datatypes.Address> list = Arrays.stream(address).map(Address::getABIAddress).collect(Collectors.toList());
+        return account.call(option, this.contractAddress, "addPrivilegeByAdmin", contractAddr.getABIAddress(),
+                new DynamicArray<>(org.web3j.abi.datatypes.Address.class, list));
     }
 
-    public String removePrivilegeByAdmin(Account.Option option, String contractAddr, String[] address) throws Exception {
-        List<Address> list = Arrays.stream(address).map(a -> new Address(a)).collect(Collectors.toList());
-        return account.call(option, contract, "removePrivilegeByAdmin", new Address(contractAddr), new DynamicArray<Address>(Address.class, list));
+    public String removePrivilegeByAdmin(Account.Option option, Address contractAddr, Address[] address) throws Exception {
+        List<org.web3j.abi.datatypes.Address> list = Arrays.stream(address).map(Address::getABIAddress).collect(Collectors.toList());
+        return account.call(option, this.contractAddress, "removePrivilegeByAdmin", contractAddr,
+                new DynamicArray<>(org.web3j.abi.datatypes.Address.class, list));
     }
 }